Core

DiffPrivacyInference.Annotation — Type

An annotation is either a sensitivity, or a privacy term.

DiffPrivacyInference.AnySSub — Type

A substitution which might be either a type- or a sensitivity substitution.

DiffPrivacyInference.Context — Type

Usually, a context is either a privacy-, or a sensitivity context.

DiffPrivacyInference.STerm — Type

This is the type of sensitivity terms. That is, formal terms for real numbers, possibly containing free variables.

It is implemented as a wrapper around the Term-type of SymbolicUtils, but also allows a term to be either simply a single symbol, or an evaluated number.

DiffPrivacyInference.Sensitivity — Type

We may use Sensitivity instead of STerm.

DiffPrivacyInference.∞ — Constant

We introduce a symbol for infinite sensitivity values. This are implemented as a free variable with the name ∞

DiffPrivacyInference.ConstrAbstr — Type

Alternative name for Constr

DiffPrivacyInference.Constraints — Type

The type of constraints is simply a list of individual constraints.

DiffPrivacyInference.ConstraintsAbstr — Type

Alternative name for Constraints

DiffPrivacyInference.DeltaNames — Type

We sometimes want to track differences of sensitivity, and type metavariable contexts. A value (ΔS,ΔT) :: DeltaNames is meant to express that ΔS and ΔT are the sensitivity, and type variable sets which were added/substracted.

DiffPrivacyInference.Full — Type

We usually carry around not only a context Γ, but additionaly a sensitivity variable context S, and a type variable context T, which track all free type/sens variables occuring in Γ. And additionally also a vector of constraints C, which express all constraints which have to hold for those free variables.

An element of type Full, i.e., a 'full context' is a combination of all of these, that is: (S,T,C,Γ) :: Full{TypeCtx}

Note: we accept different kinds of contexts here, the one intended is stated in curly braces after Full.

See also: TypeContext, PrivacyContext, SensitivityContext

DiffPrivacyInference.NameCtx — Type

A container for keeping track of active metavariables. See also: addNewName

DiffPrivacyInference.Privacy — Type

A privacy term is a tuple (ϵ,δ) of two formal expressions, here simply implemented as a pair of sensitivity terms.

DiffPrivacyInference.PrivacyContext — Type

A privacy context assigns not only a type, but also a privacy term to every variable.

DiffPrivacyInference.SSSub — Type

A single sensitivity substitution, e.g. (x, η) means x := η

DiffPrivacyInference.SVarCtx — Type

A context of sensitivity metavariables is simply a name context.

DiffPrivacyInference.SensitivityContext — Type

A sensitivity context assigns not only a type, but also a sensitivity term to every variable.

DiffPrivacyInference.Substitutions — Type

A list of multiple substitutions (of any kind).

DiffPrivacyInference.TSSub — Type

A single type substitution, e.g. (x, τ) means x := τ

DiffPrivacyInference.TVarCtx — Type

A context of type metavariables is simply a name context.

DiffPrivacyInference.TypeContext — Type

A simple context, assigning to a variable a type.

Base.isequal — Method

We override the equality on name contexts to be comparing by value instead of by reference.

Base.isequal — Method

We override the equality on constraints to be comparing by value instead of by reference.

Base.isequal — Method

We override equality of type operations to be by value instead of by reference.

Base.show — Method

Pretty printing for sensitivity contexts.

Base.show — Method

Pretty printing for a single constraint.

Base.show — Method

Pretty printing for full contexts.

Base.show — Method

Pretty printing for constraints.

DiffPrivacyInference.addNewName — Method

addNewName(N::NameCtx,hint::Symbol) :: Tuple{NameCtx,Symbol}

Requests a fresh metavariable name from N. The hint is used as a basis from which to generate the name. Returns it, as well as a modified instance of N, which should be the only one used from now on.

DiffPrivacyInference.create_DMType — Method

Make a proper DMType out of τ, adding sensitivity and type variables to S and T if necessary.

DiffPrivacyInference.emptyConstraints — Method

Creates an empty set of constraints.

DiffPrivacyInference.emptySVarCtx — Method

Creates an empty sensitivity context.

DiffPrivacyInference.emptyTVarCtx — Method

Creates an empty type context.

DiffPrivacyInference.free_SVars — Method

free_SVars(t :: DMType) :: Vector{Symbol}

Computes the free sensitivity variables in a type t.

DiffPrivacyInference.free_SVars — Method

free_SVars(c :: Constr) :: Vector{Symbol}

Computes the free sens variables in a constraint c.

DiffPrivacyInference.free_SVars — Method

Computes the free sensitivity variables in an assignment.

DiffPrivacyInference.free_SVars — Method

free_SVars(s :: Sensitivity) :: Vector{Symbol}

Computes the free sensitivity variables in a sensitivity term.

DiffPrivacyInference.free_SVars — Method

free_SVars(s :: SymbolOrSens) :: Vector{Symbol}

Computes the free sensitivity variables in a sensitivity term, which also might just be a symbol.

DiffPrivacyInference.free_SVars — Method

free_SVars(c :: Constr) :: Vector{Symbol}

Computes the free sens variables in a list of constraint c.

DiffPrivacyInference.free_TVars — Method

free_TVars(t :: DMType) :: Vector{Symbol}

Computes the free type variables in a type t.

DiffPrivacyInference.free_TVars — Method

free_TVars(c :: Constr) :: Vector{Symbol}

Computes the free type variables in a constraint c.

DiffPrivacyInference.free_TVars — Method

Computes the free sensitivity variables in an assignment.

DiffPrivacyInference.free_TVars — Method

free_TVars(s :: Sensitivity) :: Vector{Symbol}

Computes the free sensitivity variables in a type. Always returns an empty vector.

DiffPrivacyInference.free_TVars — Method

free_SVars(s :: SymbolOrSens) :: Vector{Symbol}

Computes the free type variables in a sensitivity term, which also might just be a symbol. Always returns the empty vector.

DiffPrivacyInference.free_TVars — Method

free_TVars(c :: Constr) :: Vector{Symbol}

Computes the free type variables in a list of constraint c.

DiffPrivacyInference.free_symbols — Method

free_symbols(ex::STerm)

Computes the free variables of the sensitivity term ex.

DiffPrivacyInference.getDMOp — Method

Get a map from some argument DMTypes to the DMTypeOp corresponding to the input julia function.

DiffPrivacyInference.prettyString — Method

Pretty printing binary type operations.

DiffPrivacyInference.prettyString — Method

Pretty printing ternary type operations.

DiffPrivacyInference.prettyString — Method

Pretty printing unary type operations.

DiffPrivacyInference.pretty_print — Method

Pretty printing for full contexts.

DiffPrivacyInference.showPretty — Method

Pretty printing applied type operations

DiffPrivacyInference.symbols — Method

Given a julia symbol x, we create a sensitivity term which simply contains this single variable.

DiffPrivacyInference.check_not_constant — Method

Check whether τ contains any non-constant types.

DiffPrivacyInference.signature — Function

signature(STCΣ :: Full{SCtx}, top::DMTypeOp)

Return the sensitivity signature of top, if it is clear yet.

If the types of top are not yet sufficiently specified to determine the signature, return nothing. Else, return the signature as specified in the Duet paper, as well as the return type of the operation and the new context with new constraints in case the op required any.

DiffPrivacyInference.apply_subs — Method

Apply a list of substitutions (which might be either type or sensitivity substitutions) to a value X of type A, for which these substitutions are defined.

DiffPrivacyInference.distribute_substitute — Method

Distribute substitutions over a vector of values. This calls recursively the function apply_subs.

DiffPrivacyInference.printSub — Method

Pretty printing of a list of substitutions.

DiffPrivacyInference.singleSSub — Method

Apply the single sensitivity substition σ to the type T. This is necessary, since types may contain sensitivity terms.

DiffPrivacyInference.singleSSub — Method

Apply a single sensitivity substitution σ to a type context Γ, by substituting all types in this context with σ.

DiffPrivacyInference.singleSSub — Method

Apply a single sensitivity substitution σ to a sensitivity context Σ, by substituting all types in this context with σ.

DiffPrivacyInference.singleSSub — Method

Apply the single sensitivity substitution σ to the constraint c.

DiffPrivacyInference.singleSSub — Method

Apply the single sensitivity substitution σ to the type operation op.

DiffPrivacyInference.singleSSub — Method

Remove the sens variable x from the context S containing active sens variables. This function exists out of correctness purposes. We always carry the contexts T and S of type- and sens-metavariables, and when substituting a meta-variable, we remove it from T, resp. S. This allows us to catch errors where a substitution is applied multiple times, by throwing an error here, if the x is not element of the currently active names (that is, was already substituted once).

DiffPrivacyInference.singleSSub — Method

Substitute in a value of type Nothing. This does nothing.

DiffPrivacyInference.singleSSub — Method

Apply the single sensitivity substitution (x := η) the sens variable a. That is, if a is exactly x, we replace it by η. Else, we do nothing.

DiffPrivacyInference.singleSSub — Method

Replace every occurence of the sens variable X by the term η in the term s. That is, compute the substitution s[X := η]. Since s or η might be already fully evaluated numbers, we simplify the resulting term in those cases as much as possible.

DiffPrivacyInference.singleSSub — Method

Apply the single sensitivity substitution σ to the list of constraint C.

DiffPrivacyInference.singleSSub — Method

Apply a single sensitivity substitution σ to a full context (S,T,C,Γ), by applying it to all parts.

DiffPrivacyInference.singleTSub — Method

Replace every occurence of the type variable X by the type ξ in the type τ. That is, compute the substitution τ[X := ξ].

DiffPrivacyInference.singleTSub — Method

Apply a single type substitution σ to a type context Γ, by substituting all types in this context with σ.

DiffPrivacyInference.singleTSub — Method

Apply a single type substitution σ to a sensitivity context Γ, by substituting all types in this context with σ.

DiffPrivacyInference.singleTSub — Method

Apply the single type substitution σ to the constraint c.

DiffPrivacyInference.singleTSub — Method

Apply the single type substitution σ to the type operation op.

DiffPrivacyInference.singleTSub — Method

Remove the type variable x from the context T containing active type variables. This function exists out of correctness purposes. We always carry the contexts T and S of type- and sens-metavariables, and when substituting a meta-variable, we remove it from T, resp. S. This allows us to catch errors where a substitution is applied multiple times, by throwing an error here, if the x is not element of the currently active names (that is, was already substituted once).

DiffPrivacyInference.singleTSub — Method

Apply a type substitution to a value of type Nothing. This does nothing.

DiffPrivacyInference.singleTSub — Method

Apply the single type substitution (x := ξ) the type variable a. That is, if a is exactly x, we replace it by ξ. Else, we do nothing.

DiffPrivacyInference.singleTSub — Method

Apply a type substitution to a sensitivity term. Since sensitivities do not contain type variables, this does nothing.

DiffPrivacyInference.singleTSub — Method

Apply the single type substitution σ to the list of constraints C.

DiffPrivacyInference.singleTSub — Method

Apply a single type substitution σ to a full context (S,T,C,Γ), by applying it to all parts.

DiffPrivacyInference.Infty — Type

A special struct that try_destructure_sensitivity will return upon encountering the ∞ symbol.

DiffPrivacyInference.substitutions_to_constraints — Method

Convert substitutions to isEqual or isEqualType constraints.

DiffPrivacyInference.try_destructure_sensitivity — Method

If s is a single symbol, infinity, or a number, return that. If not, return nothing.

DiffPrivacyInference.unify_DMType — Method

unify_DMType(τ :: DMType, ρ :: DMType)

Unify the types τ and ρ, returning the new type as well as all constraints added and substitutions made in the process.

DiffPrivacyInference.unify_Sensitivity — Method

Unify s1 and s2 by attempting to destructure or adding an equality constraint upon failure.

DiffPrivacyInference.add — Method

Add all Σs, unifying types of variables where the contexts disagree of the type.

DiffPrivacyInference.add_TypeOp — Method

add_TypeOp((S,T,C), op :: DMTypeOp)

Add op to the constraints C, also creating the necessary type and sensitivity variables. Return the new context, the DMType of the result of the operation, and its signature.

DiffPrivacyInference.merge_contexts — Method

merge_contexts(combine::Function, S, T, C, Σ1, Σ2)

Make a new context that has all variables of both input contexts. Unify types if the contexts disagree on the type of a variable. Apply the combine function on the annotations of the variables present in both contexts.

DiffPrivacyInference.merge_contexts — Method

merge_contexts(combine::Function, S, T, C, Σs::Array)

Make a new context that has all variables of all input contexts. Unify types if the contexts disagree on the type of a variable. Apply the combine function on the annotations of the variables present in multiple contexts.

DiffPrivacyInference.remove — Method

Return new context with the given entries removed.

DiffPrivacyInference.scale — Method

Scale all sensitivities in Σ by r.

DiffPrivacyInference.scale — Method

Scale all sensitivities in Σ by r.

DiffPrivacyInference.scale — Method

Scale all sensitivities in Σ by r.

DiffPrivacyInference.upperbound — Method

Combine Σs using the maximum of annotations, unifying types of variables where the contexts disagree of the type.

DiffPrivacyInference.add_Cs — Method

Add the given constraints to the monad's constraint list.

DiffPrivacyInference.add_op — Method

Add a DMTypeOp constraint for the nargs-ary operation accoding to opf.

DiffPrivacyInference.add_svar — Method

Add a newly created sensitivity variable to the monad's sensitivity variable context S and return it.

DiffPrivacyInference.add_type — Method

addtype(maketype::Function) :: TC

Add a newly created typevar to the monad's type variable context T, made by calling make_type::TVarCtx => TVarCtx x DMType. Return the new type.

DiffPrivacyInference.extract_Cs — Method

Return the current constraint list.

DiffPrivacyInference.lookup_var_sens — Method

Return variable x's current sensitivity.

DiffPrivacyInference.lookup_var_type — Method

Return variable x's current type.

DiffPrivacyInference.mcreate_DMType — Method

Convert a given julia type into a DMType and return it.

DiffPrivacyInference.mscale — Method

Scale a TC Monad's sensitivity context by s.

DiffPrivacyInference.remove_var — Method

Delete x from the current context.

DiffPrivacyInference.set_var — Method

Set sensitivity of x to s and type to τ.

DiffPrivacyInference.subtype_of — Method

Add a subtype constraint τ1 ⊑ τ2 to the monad's constraint list.

DiffPrivacyInference.unify — Method

Unify DMTypes τ1 and τ2 within a TC monad.