Public interface

DiffPrivacyInference.build_tc — Method

Construct a TC monad containing the computation of inferring t's sensitivity.

DiffPrivacyInference.infer_sensitivity — Method

Infer the type of a DMTerm.

DiffPrivacyInference.infer_sensitivity_from_file — Method

infersensitivityfrom_file(s::String) :: DMType

Given a filename pointing to some julia code, infer the type of the corresponding expression. includes are resolved and parsed as well. The return value is the type of the expression, a DMType, our custom types defined in definitions.jl. If the expression is a function, it should be an Arr type with the argument sensitivities encoded next to the argument types.

DiffPrivacyInference.infer_sensitivity_from_string — Method

infersensitivityfrom_string(s::String) :: DMType

Given a String of julia code, infer the type of the corresponding expression. The return value is the type of the expression, a DMType, our custom types defined in definitions.jl. If the expression is a function, it should be an Arr type with the argument sensitivities encoded next to the argument types.

Examples

julia> t = infersensitivityfromstring(" function test(x, y) f(x) = 500(x + y) z = 1 g(x) = zx z = 2/5 f(g(x)) end ") prettyprint(t) Arr(Tuple{Union{Number, SymbolicUtils.Symbolic, var"#s18"} where var"#s18"<:SymbolicUtils.Sym,DMType}[(200.0, TVar(:oparg3)), (500, TVar(:oparg17))], TVar(:sup_29))

julia> prettyprint(t) "(tvar.oparg3 @(200.0), tvar.oparg17 @(500)) ==> tvar.sup29" ``` The output means that the expression is a two-argument function that is 200-sensitive in the first and 500-sensitive in the second argument. The first argument is an Integer, and the second arguments' and the return type could not be inferred (that's what the TVar means).