Typechecking

DiffPrivacyInference.ArithmeticsError — Type

An error to throw upon invalid arithmetic operations.

DiffPrivacyInference.ConstraintViolation — Type

An error to throw upon finding a constraint that is violated.

DiffPrivacyInference.NoChoiceFound — Type

An error to throw upon absence of a matching method in multiple dispatch.

DiffPrivacyInference.NotInScope — Type

An error to throw upon encountering a variable that was never defined.

DiffPrivacyInference.NotNumeric — Type

An error to throw upon using a non-numeric value in a numeric context.

DiffPrivacyInference.NotSubtype — Type

An error to throw upon vioaltion of a sunbype constraint.

DiffPrivacyInference.NotSupremum — Type

An error to throw upon violation of a supremum constraint.

DiffPrivacyInference.UnificationError — Type

An error to throw upon a failed type or sensitivity unification attempt.

DiffPrivacyInference.WrongNoOfArgs — Type

An error to throw upon calling a function with the wrong number of arguments.

DiffPrivacyInference.apply_subs — Method

Apply all substitutions encoded in the constraints of the TC monad m to the DMType τ.

DiffPrivacyInference.check_term — Method

check_term(t::DMTerm, scope = Dict{Symbol, Vector{DMTerm}}()) :: TC

Typecheck the input DMTerm and return the resulting computation as a TC monad. The result will have a lot of unresolved constraints.

DiffPrivacyInference.choice_could_match — Method

See if a call with argument types args would fit a method with signature cs, if the type variables in args would resolve to the right thing.

DiffPrivacyInference.keep_least_general — Method

Remove entries from cs that are supertypes of some other entry.

DiffPrivacyInference.msimplify_constraints — Method

simplify_constraints() :: TC{Tuple{}}

Evaluate all constraints that can be simplified.

DiffPrivacyInference.simplify_constraints_lose_generality — Method

simplifyconstraintslose_generality() :: TC

Simplify as many constraints as possible, then assume all free type variables involved in DMTypeOps are non-constant and simplify again.

This worsens the sensitivity, as non-constant types yield larger sensitivity penalty. It is hence legal to do, but should only be done if no other simplification is possible for best results.

DiffPrivacyInference.try_eval_isSubtypeOf — Method

eval_arr(Φ :: Full{A}, τ1 :: DMType, τ2 :: DMType)

Attempt to simplify a subtyping constraint τ1 ⊑ τ2, potentially adding new variables and constraints to the context. Return nothing if there is not enough information to simplify, and throw an error if the constraint is violated.

DiffPrivacyInference.try_get_direct_supertypes — Method

Get the list of direct supertypes of τ, or return nothing if τ has type variables.