Builtins

DiffPrivacyInference.DMGrads — Type

A wrapper for Zygote.Grads, so we can control that only typecheckable operations are executed on the gradient.

Examples

A black-box function computing the gradient of some DMModel, given a loss function loss:

function unbounded_gradient(model::DMModel, d::Vector, l) :: BlackBox()
   gs = Flux.gradient(Flux.params(model.model)) do
           loss(d,l,model)
        end
   return DMGrads(gs)
end

DiffPrivacyInference.DMModel — Type

A wrapper for Flux models, so we can control that only typecheckable operations are executed on the model. What you put inside this wrapper needs to at least support calling Flux.params on it.

Examples

Intialize a Flux neural network:

 DMModel(Flux.Chain(
         Flux.Dense(28*28,40, Flux.relu),
         Flux.Dense(40, 10),
         Flux.softmax))

Note that construction of models cannot be typechecked and needs to happen inside black-box functions that return the model. So a typecheckable function could look like this:

function init_model() :: BlackBox()
   DMModel(Flux.Chain(
           Flux.Dense(28*28,40, Flux.relu),
           Flux.Dense(40, 10),
           Flux.softmax))
end

DiffPrivacyInference.Data — Type

Annotation for real numbers with the discrete metric, i.e. d(a,b) = (a==b) ? 1 : 0 Use it to tell the typechecker you want to infer sensitivity/privacy of a function variable w.r.t. to the discrete metric. An alias for julia's Real type, so you cannot dispatch on it.

DiffPrivacyInference.PrivacyFunction — Type

Annotation for variables of a function that are privacy functions themselves. You have to annotate privacy function function arguments, otherwise typechecking will assume a non-private function and fail if you insert a privacy function.

DiffPrivacyInference.BlackBox — Method

Annotation for functions that cannot be typechecked. Their arguments will be assigned infinite sensitivity. Note that it is not allowed to mutate any of the arguments in a function like this, if you do the typechecking result will be invalid!

Examples

A function calling an imported qualified name, which is not permissible in non-black-boxes:

loss(X, y, m::DMModel) :: BlackBox() = Flux.crossentropy(m.model(X), y)

DiffPrivacyInference.MetricGradient — Method

MetricGradient(T, N<:Norm)

Annotate gradients with the desired metric you want them to be measured in by the typechecker. Just maps to DMGrad.

DiffPrivacyInference.MetricMatrix — Method

MetricMatrix(T, N<:Norm)

Annotate matrices with the desired metric you want them to be measured in by the typechecker. Just maps to Matrix{T}.

DiffPrivacyInference.MetricVector — Method

MetricVector(T, N<:Norm)

Annotate matrices with the desired metric you want them to be measured in by the typechecker. Just maps to Vector{T}.

DiffPrivacyInference.Priv — Method

Annotation for functions whose differential privacy we want to infer.

Examples

A privacy function with argument x whose privacy will be inferred and argument y of type Integer whose privacy we're not interested in:

function foo(x, y::Static(Integer)) :: Priv()
   x
end

DiffPrivacyInference.Static — Method

Annotation for function arguments whose privacy is of no interest to us. Their privacy will most likely be set to infinity to allow tighter bounds on other arguments.

Examples

A privacy function with argument x whose privacy will be inferred and argument y of type Integer whose privacy we're not interested in:

function foo(x, y::Static(Integer)) :: Priv()
   x
end

DiffPrivacyInference.above_threshold — Method

above_threshold(queries :: Vector{Function}, epsilon :: Real, d, T :: Number) :: Integeri

The above-threshold mechanism. Input is a vector of 1-sensitive queries on dataset d mapping to the reals. Returns the index of the first query whose result at d plus (4/epsilon)-Laplacian noise is above the given threshold T plus (2/epsilon)-Laplacian noise. This is (epsilon,0)-private in d!

DiffPrivacyInference.clip! — Method

clip!(l::Norm, g::DMGrads) :: Nothing

Clip the gradient, i.e. scale by 1/norm(g) if norm(g) > 1. Mutates the gradient, returns nothing.

DiffPrivacyInference.clip — Method

clip(l::Norm, g::AbstractVector)

Return a clipped copy of the input vector, i.e. scale by 1/norm(g) if norm(g) > 1.

DiffPrivacyInference.clip — Method

clip(l::Norm, g::DMGrads) :: Nothing

Return a clipped copy of the gradient, i.e. scale by 1/norm(g) if norm(g) > 1.

DiffPrivacyInference.clip — Method

clip(v::T, upper::T, lower::T) where T <: Number

Clip the number v, i.e. return v if it is in [lower,upper], return upper if v is larger than upper, and return lower if v is smaller than lower.

DiffPrivacyInference.clone — Method

clone(g::DMGrads)

Create and return a copy of a DMGrads object, where only the gradient part of the Zygote gradient is copied while the part pointing to the parameters of a model is kept. Thus we get an object that we can mutate safely while retaining information on which entry of the gradient belongs to which parameter of which model. If you want to return a DMGrads object from a function, you have to return a copy.

Examples

A function returning a copy of the gradient object:

function compute_and_scale_gradient(model::DMModel, d, l) :: BlackBox()
   gs = unbounded_gradient(model, d, l)
   scale_gradient!(100, gs)
   return clone(gs)
end

DiffPrivacyInference.disc — Method

disc(n::Real) :: Data

Return n, but let the typechecker know that you want it to be measured in the discrete metric.

DiffPrivacyInference.exponential_mechanism — Method

exponential_mechanism(r::Number, eps::Number, xs::Vector, u::Function)

Return an element of the input vector xs based on the score given by the function u, mapping from the elements of xs to a real number. The probability for element e to be chosen is proportional to exp(eps*u(e)/(2*r)). The mechanism is (eps,0)-private in the variables that u is r-sensitive in.

DiffPrivacyInference.fold — Method

fold(f::Function, i, m::AbstractMatrix)

Fold the function f over all entries of m, using initial value i.

DiffPrivacyInference.gaussian_mechanism! — Method

gaussian_mechanism!(s::Real, ϵ::Real, δ::Real, g::DMGrads) :: Nothing

Apply the gaussian mechanism to the input gradient, adding gaussian noise with SD of (2 * log(1.25/δ) * s^2) / ϵ^2) to each gradient entry seperately. This introduces (ϵ, δ)-differential privacy to all variables the gradient depends on with sensitivity at most s. Mutates the gradient, returns nothing.

DiffPrivacyInference.gaussian_mechanism — Method

gaussian_mechanism(s::Real, ϵ::Real, δ::Real, g)

Apply the gaussian mechanism to the input, adding gaussian noise with SD of (2 * log(1.25/δ) * s^2) / ϵ^2). This introduces (ϵ, δ)-differential privacy to all variables the input depends on with sensitivity at most s. Makes a copy of the input and returns the noised copy.

DiffPrivacyInference.laplacian_mechanism! — Method

laplacian_mechanism!(s::Real, ϵ::Real, g::DMGrads) :: Nothing

Apply the laplacian mechanism to the input, adding laplacian noise with scaling parameter of (s / ϵ) and location zero to each gradient entry seperately. This introduces (ϵ, 0)-differential privacy to all variables the input depends on with sensitivity at most s. Mutates the input, returns nothing.

DiffPrivacyInference.laplacian_mechanism — Method

laplacian_mechanism(s::Real, ϵ::Real, g)

Apply the laplacian mechanism to the input, adding laplacian noise with scaling parameter of (s / ϵ) and location zero to each gradient entry seperately. This introduces (ϵ, 0)-differential privacy to all variables the input depends on with sensitivity at most s. Makes a copy of the input, then noises and returns the copy.

DiffPrivacyInference.map_cols — Method

map_cols(f::Function, m::AbstractMatrix)

Map the Vector-to-Vector-function f to the columns of m.

DiffPrivacyInference.map_cols_binary — Method

map_cols_binary(f::Function, m::AbstractMatrix, n::AbstractMatrix)

Map the binary Vector-to-Vector-function f to the columns of m and n.

DiffPrivacyInference.map_rows — Method

map_rows(f::Function, m::AbstractMatrix)

Map the Vector-to-Vector function f to the rows of m.

DiffPrivacyInference.map_rows_binary — Method

map_rows_binary(f::Function, m::AbstractMatrix, n::AbstractMatrix)

Map the binary Vector-to-Vector-function f to the columns of m and n.

DiffPrivacyInference.norm_convert! — Method

norm_convert!(m::T) :: T

Make a clipped vector/gradient measured using the discrete metric into a vector/gradient measured with the clipping norm instead. Does not change the value of the argument. It can be used to enable using a gradient obtained from a black box computation (hence being in discrete-norm land) to be put into e.g. the gaussian mechanism (which expects the input to be in L2-norm land).

DiffPrivacyInference.norm_convert — Method

norm_convert(m::T) :: T

Make a clipped vector/gradient measured using the discrete norm into a vector/gradient measured with the clipping norm instead. Does not change the value of the argument. It can be used to enable using a gradient obtained from a black box computation (hence being in discrete-norm land) to be put into e.g. the gaussian mechanism (which expects the input to be in L2-norm land).

DiffPrivacyInference.parallel_private_fold_rows! — Method

parallel_private_fold_rows(f::Function, i, m::AbstractMatrix, n::AbstractMatrix)

Fold the privacy function f :: Vector -> Vector -> I -> I over the two input matrices' rows simultaneously. Allows for f to mutate the accumulator, returns nothing. This is parallel composition on the rows of m and n, so if f is (eps,del)-private in it's first two arguments, the fold is (eps,del)-private in the input matrices. The input matrices are expected to be measured in the discrete norm.

DiffPrivacyInference.parallel_private_fold_rows — Method

parallel_private_fold_rows(f::Function, i, m::AbstractMatrix, n::AbstractMatrix)

Fold the privacy function f :: Vector -> Vector -> I -> I over the two input matrices' rows simultaneously. This is parallel composition on the rows of m and n, so if f is (eps,del)-private in it's first two arguments, the fold is (eps,del)-private in the input matrices. The input matrices are expected to be measured in the discrete norm.

DiffPrivacyInference.reduce_cols — Method

reduce_cols(f::Function, m::AbstractMatrix)

Apply the privacy function f :: (r x 1)-Matrix -> T to each column of the (r x c)-Matrix m, return a vector of the results. If f is (eps,del)-private in its argument, the reduction is (r*eps, r*del)-private in m.

DiffPrivacyInference.row_to_vec — Method

row_to_vec(m::AbstractMatrix) :: Vector

Make the one-row matrix m into a vector.

DiffPrivacyInference.sample — Method

sample(n::Integer, m::AbstractMatrix, v::AbstractMatrix) :: Tuple

Take a uniform sample (with replacement) of n rows of the matrix m and corresponding rows of matrix v.

DiffPrivacyInference.scale_gradient! — Method

scale_gradient!(s::Number, gs::DMGrads) :: Nothing

Scale the gradient represented by the Zygote.Grads struct wrapped in the input DMGrads gs by the scalar s. Mutates the gradient, returs nothing.

DiffPrivacyInference.subtract_gradient! — Method

subtract_gradient!(m::DMModel, gs::DMGrads) :: Nothing

Subtract the gradient represented by the Zygote.Grads struct wrapped in the input DMGrads gs from the parameters of the model m. Mutates the model, returns nothing.

DiffPrivacyInference.sum_gradients — Method

sum_gradients(g::DMGrads, gs::DMGrads...) :: DMGrads

Sum two or more DMGrads gradients. Errors if they belong to different DMModels.

DiffPrivacyInference.vec_to_row — Method

vec_to_row(v::AbstractVector) :: Matrix

Make the vector v into a one-row matrix.

DiffPrivacyInference.zero_gradient — Method

zero_gradient(m::DMModel) :: DMGrads

Create a zero gradient for the given model.